requirements.txt
databases==0.9.0
pydantic==2.6.4
pydantic_settings==2.2.1
python-dotenv==1.0.1
Requests==2.31.0
requests_oauthlib==2.0.0
rich==13.7.1
SQLAlchemy==2.0.29
uvicorn==0.29.0
wxcadm==4.2.6
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
dnspython==2.6.1
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
meraki==1.27.0
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
psycopg2==2.9.9
pydantic_core==2.16.3
Pygments==2.17.2
pyhumps==3.8.0
requests-toolbelt==1.0.0
sniffio==1.3.1
srvlookup==3.0.0
starlette==0.36.3
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
xmltodict==0.13.0
yarl==1.9.4
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
shellingham==1.5.4
typer==0.12.3
typing_extensions==4.11.0
